ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.3",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.16",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
